% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_interface.R
\name{datafile_filter}
\alias{datafile_filter}
\title{Data file filter}
\usage{
datafile_filter(filelist, filterby)
}
\arguments{
\item{filelist}{The return data, a vector of characters, from sample_files}

\item{filterby}{A string, naming a platform}
}
\value{
List of files
}
\description{
Filter the results from datafilenamekeys
}
\details{
The call to the datafilenamekeys API returns a list with platform types
in the strings. This filters down the list of files by platform.
}
\examples{
\dontrun{
  fs <- sample_files("TCGA-A7-A6VV-10A")
  datafile_filter(sf, ".bam")
 }
}

