% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_interface.R
\name{datafiles_from_cohort}
\alias{datafiles_from_cohort}
\title{Data files from a cohort}
\usage{
datafiles_from_cohort(cohort_id, pipeline = NULL, platform = NULL, a_token)
}
\arguments{
\item{cohort_id}{The list of filter parameters defining the cohort}

\item{pipeline}{The pipeline used in data generation}

\item{platform}{The platform used for data generation}

\item{token}{Your auth token}
}
\description{
Using a defined cohort, get file paths to a Google bucket
}
\examples{
\dontrun{
  preview_cohort(list(Study="BRCA"))
}
}

