% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_interface.R
\name{list_cohorts}
\alias{list_cohorts}
\title{List cohorts}
\usage{
list_cohorts(a_token)
}
\arguments{
\item{a_token}{Auth token as returned by isb_init()}
}
\description{
Using your auth token, as generated by isb_init(), get a list of previously created cohorts
}
\examples{
\dontrun{
  list_cohorts(mytoken)
 }
}
