% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_interface.R
\name{patient_details}
\alias{patient_details}
\title{Get patient details}
\usage{
patient_details(patient_barcode)
}
\arguments{
\item{sample_barcode}{A TCGA patient barcode like "TCGA-02-0001"}
}
\description{
Get details on a particular patient, given a sample barcode
}
\examples{
\dontrun{
  patient_details("TCGA-A7-A6VV")
 }
}

