% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_interface.R
\name{preview_cohort}
\alias{preview_cohort}
\title{Preview cohorts}
\usage{
preview_cohort(filter_list)
}
\arguments{
\item{filter_list}{The list of filter parameters defining the cohort}
}
\description{
Using your auth token, create a cohort, without saving
}
\examples{
\dontrun{
  preview_cohort(list(Study="BRCA"))
}
}

