% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_interface.R
\name{sample_details}
\alias{sample_details}
\title{Get sample details}
\usage{
sample_details(sample_barcode)
}
\arguments{
\item{sample_barcode}{A TCGA sample barcode}
}
\description{
Get details on a particular sample, given a sample barcode
}
\examples{
\dontrun{
  sample_details("TCGA-A7-A6VV-10A")
}
}

