% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_interface.R
\name{sample_files}
\alias{sample_files}
\title{Get sample files}
\usage{
sample_files(sample_barcode)
}
\arguments{
\item{sample_barcode}{A TCGA barcode.}
}
\value{
List of files across platforms.
}
\description{
Get file locations using sample barcode
}
\details{
Uses the API to get a list of files for a given barcode.
}
\examples{
\dontrun{
  fs <- sample_files("TCGA-A7-A6VV-10A")
 }
}

