% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_interface.R
\name{save_cohorts}
\alias{save_cohorts}
\title{Save cohorts}
\usage{
save_cohorts(a_token, cohort_name, filter_list)
}
\arguments{
\item{a_token}{Your auth token}

\item{cohort_name}{The name of the new cohort}

\item{filter_list}{The list of filter parameters defining the cohort}
}
\description{
Using your auth token, save a new cohort
}
\examples{
\dontrun{
  save_cohorts(mytoken, "new_cohort", list(list(Study="BRCA")))
}
}

