% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_interface.R
\name{list_cohorts}
\alias{list_cohorts}
\title{List cohorts}
\usage{
list_cohorts(a_token)
}
\arguments{
\item{sample_barcode}{A TCGA patient barcode like "TCGA-02-0001"}
}
\description{
Using your auth token, get a list of saved cohorts
}
\examples{
\dontrun{
  my
  list_cohorts(mytoken)
 }
}

