% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_interface.R
\name{list_cohorts}
\alias{list_cohorts}
\title{List cohorts}
\usage{
list_cohorts(a_token)
}
\arguments{
\item{a_token}{Auth token as returned by isb_init()}
}
\description{
Using your auth token, as generated by isb_init(), get a list of previously created cohorts.
The result is a list, with two important elements: count (giving the number of saved cohorts) and items
(which contain the cohort ID as result$items[[x]]$id, and name as result$items[[x]]$name).
}
\examples{
\dontrun{
  list_cohorts(mytoken)
 }
}
